clear app state on logout

+++ src/store/reducers/ormReducer.js
@@ -1,4 +1,4 @@
-import { LOGOUT, CLEAR_CURRENT_USER } from '../../components/Login/actions'
+import { LOGOUT } from '../../components/Login/actions'
 import { SIGNUP, ADD_SKILL, REMOVE_SKILL, UPDATE_USER_SETTINGS_PENDING } from '../../components/SignupFlow/SignupFlow.store'
 import {
   CREATE_COMMENT
@@ -25,10 +25,11 @@ export default function ormReducer (state = {}, action) {
 
   switch (type) {
     case LOGOUT:
-    case CLEAR_CURRENT_USER:
       me = session.Me.first()
-      me.memberships.delete()
-      me.delete()
+      if (me) {
+        me.memberships.delete()
+        me.delete()
+      }
       break
 
     case CREATE_COMMENT:
